+++ test/metro/js/metro.js
@@ -3602,7 +3602,7 @@ var isTouch = (('ontouchstart' in window) || (navigator["MaxTouchPoints"] > 0) |
 var Metro = {
 
     version: "4.3.2",
-    compileTime: "13/10/2019 19:55:29",
+    compileTime: "15/10/2019 18:21:22",
     buildNumber: "739",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
@@ -29127,7 +29127,8 @@ var TreeView = {
         checks = check.closest("li").find("ul input[type=checkbox]");
         checks.attr("data-indeterminate", false);
         checks.prop("checked", checked);
-
+        checks.trigger('change');
+        
         all_checks = [];
 
         $.each(element.find("input[type=checkbox]"), function(){
@@ -29144,6 +29145,7 @@ var TreeView = {
             if (children > 0 && children_checked === 0) {
                 ch.attr("data-indeterminate", false);
                 ch.prop("checked", false);
+                ch.trigger('change');
             }
 
             if (children_checked === 0) {
@@ -29154,6 +29156,7 @@ var TreeView = {
                 } else if (children === children_checked) {
                     ch.attr("data-indeterminate", false);
                     ch.prop("checked", true);
+                    ch.trigger('change');
                 }
             }
         });
@@ -29313,6 +29316,7 @@ var TreeView = {
 };
 
 Metro.plugin('treeview', TreeView);
+
 
 var ValidatorFuncs = {
     required: function(val){